[REFACTOR] 알림 권한 요청을 위해 만들었던 확장함수들을 별도 파일로 분리했어요.
